requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==4.2.14
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
